add redirect when user isn't authentication

+++ posts/views.py
@@ -4,6 +4,7 @@ from .models import Post, LikeDislike
 from django.views.generic.base import RedirectView
 from django.contrib.auth.models import User
 from django.views.generic.list import ListView
+from django.shortcuts import redirect
 import json
 
 
@@ -30,36 +31,38 @@ class PostListView(ListView):
         # likes_count = LikeDislike.object.getLikes(post_id=post_id)
         # show dislike count
         if request.method == 'POST':
-            if request.POST.get('action'):
-                post_id = request.POST.get('post_id')
-                user_id = request.POST.get('user_id')
-                action = request.POST.get('action')
+            if request.user.is_authenticated:
+                if request.POST.get('action'):
+                    post_id = request.POST.get('post_id')
+                    user_id = request.POST.get('user_id')
+                    action = request.POST.get('action')
 
-                if action == 'like':
-                    LikeDislike.objects.update_or_create_like(post_id, user_id)
+                    if action == 'like':
+                        LikeDislike.objects.update_or_create_like(post_id, user_id)
 
-                elif action == 'unlike':
-                    LikeDislike.objects.delete_after_unlike(post_id, user_id)
+                    elif action == 'unlike':
+                        LikeDislike.objects.delete_after_unlike(post_id, user_id)
 
-                elif action == 'dislike':
-                    LikeDislike.objects.update_or_create_dislike(post_id, user_id)
+                    elif action == 'dislike':
+                        LikeDislike.objects.update_or_create_dislike(post_id, user_id)
 
-                elif action == 'undislike':
-                    LikeDislike.objects.delete_after_undislike(post_id, user_id)
-                else:
-                    return None
+                    elif action == 'undislike':
+                        LikeDislike.objects.delete_after_undislike(post_id, user_id)
+                    else:
+                        return None
 
-                # Post.objects.get_rating_by_ajax(post_id)
-                # store count like and dislike
-                count_likes     = LikeDislike.objects.filter(post__id=post_id, rating_action=1).count()
-                count_dislikes  = LikeDislike.objects.filter(post__id=post_id, rating_action=0).count()
+                    # Post.objects.get_rating_by_ajax(post_id)
+                    # store count like and dislike
+                    count_likes     = LikeDislike.objects.filter(post__id=post_id, rating_action=1).count()
+                    count_dislikes  = LikeDislike.objects.filter(post__id=post_id, rating_action=0).count()
 
-                # rating.dict(count_likes=count_likes, count_dislikes=count_dislikes)
-                return JsonResponse({
-                    'likes': count_likes,
-                    'dislikes': count_dislikes
-                })
-
+                    # rating.dict(count_likes=count_likes, count_dislikes=count_dislikes)
+                    return JsonResponse({
+                        'likes': count_likes,
+                        'dislikes': count_dislikes
+                    })
+            else:
+                return redirect('post_url')
 
 
     def get_context_data(self, **kwargs):
